refactor pipeline approach
