+++ unstructured/ingest/pipeline/doc_factory.py
@@ -1,0 +1,15 @@
+import typing as t
+from dataclasses import dataclass
+
+from unstructured.ingest.pipeline.interfaces import DocFactoryNode
+
+
+@dataclass
+class DocFactory(DocFactoryNode):
+    def initialize(self):
+        self.source_doc_connector.initialize()
+
+    def run(self) -> t.Iterable[str]:
+        docs = self.source_doc_connector.get_ingest_docs()
+        json_docs = [doc.to_json() for doc in docs]
+        return json_docs
